apaga componentes antigos e reorganiza
